move prepare_data and calculate_rankings to main.py
